% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itabplot.R
\name{itabplot}
\alias{itabplot}
\title{Interactive tableplot}
\usage{
itabplot(x, ..., height = NULL, width = NULL)
}
\arguments{
\item{x}{\code{data.frame} or \code{ffdf} object used for plotting.}

\item{...}{parameters that will be given to \code{tableplot}}
}
\description{
\code{itabplot} is an interactive version of \code{\link{tableplot}}. It starts
your browser and allows for zooming, panning and sorting the tableplot. This
version can be used for explorative usage, while \code{tableplot} can be used for
publication purposes.
It needs the same parameters as tabplot.
}
\seealso{
\code{\link{tableplot}}
}

